requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.2
gunicorn==19.6.0
psycopg2==2.6.1
python-decouple==3.0
django-gravatar2==1.4.0
django-notifications-hq==1.2
